fix(consensus): make compound-id allocators provably disjoint (no collision)
Pre-existing latent consensus hazard surfaced by the verify-f8-and-d5 workflow
and confirmed by direct reading. Two id allocators minted OVERLAPPING u64
compound ids within one execution:
  - CompoundIds (runtime_types.rs, fresh per exec via CompoundIds::default() at
    executor/mod.rs:42, counts 0,1,2,...) -> NEWARRAY/NEWSTRUCT/NEWMAP/PACK.
  - global atomic NEXT_STACK_ITEM_ID (vm/graph.rs, from 1, NEVER reset) ->
    NEWBUFFER, NEWMAP(collections), CONVERT, and the ABI/host boundary.
The alias machinery (replace_alias/find_affected_indices) and the refcounter
(count_references) key PURELY on the u64 id with NO type check, so when a
NEWARRAY id and a NEWBUFFER id collided, mutating one CLOBBERED the other with
type-confused contents (Buffer<->Array) and count_references dedup'd two distinct
compounds, UNDER-counting the MaxStackSize(2048) gate (a script that should FAULT
could HALT). Reachable from pure bytecode; the global atomic is also
non-deterministic across executions (fork-prone on collision).

Fix (minimal, root-cause, provably complete): tag every global-atomic id with the
high bit (1<<63). CompoundIds counts up from 0 and cannot reach 2^63 in any
single execution, so the two allocators now occupy DISJOINT bands -> a
cross-allocator collision is impossible, regardless of which call site uses which
allocator. Within an execution every compound id is therefore unique, so the
id-keyed machinery is sound and refcounts are exact.

Why this preserves determinism: compound ids are NEVER serialized (fast_codec
encodes (tag,len,items)/(tag,bytes), dropping the id field) and are only ever
compared for EQUALITY, never by value (EQUAL on Array/Map/Buffer is reference
identity; count_references/Tarjan use id equality). So the process-global,
non-deterministic low bits are consensus-invisible. Verified no id is ever
truncated to u32 or used as a Vec index. next_stack_item_id() now returns u64
(was usize; required since riscv32 usize is 32-bit); redundant  casts at
call sites dropped.

This also UNBLOCKS D-5 (Struct.Clone share-by-reference): its shared children can
no longer collide with an unrelated compound. D-5 + the PACK/PACKSTRUCT
store-by-reference fix follow next.

Full suite incl. official VMUT green; disjoint-band invariant regression test
added. GATED SOURCE: refcounter/blob-affecting -> guest-blob regen + mainnet
replay are the gates.

+++ src/abi/callback_codec.rs
@@ -155,7 +155,7 @@ pub fn decode_stack_result_into(
         TAG_OK_BUFFER => {
             stack.clear();
             stack.push(StackValue::Buffer(
-                crate::next_stack_item_id() as u64,
+                crate::next_stack_item_id(),
                 cursor.read_bytes()?.to_vec(),
             ));
             Ok(())
@@ -196,7 +196,7 @@ pub fn decode_stack_result(bytes: &[u8]) -> Result<Result<Vec<StackValue>, Strin
         TAG_OK_ITERATOR => Ok(vec![StackValue::Iterator(cursor.read_u64()?)]),
         TAG_OK_POINTER => Ok(vec![StackValue::Pointer(cursor.read_i64()?)]),
         TAG_OK_BUFFER => Ok(vec![StackValue::Buffer(
-            crate::next_stack_item_id() as u64,
+            crate::next_stack_item_id(),
             cursor.read_bytes()?.to_vec(),
         )]),
         _ => return Err("invalid stack result tag".to_string()),
@@ -285,7 +285,7 @@ fn decode_stack_value_depth(cursor: &mut Cursor<'_>, depth: usize) -> Result<Sta
             for _ in 0..len {
                 items.push(decode_stack_value_depth(cursor, depth + 1)?);
             }
-            Ok(StackValue::Array(crate::next_stack_item_id() as u64, items))
+            Ok(StackValue::Array(crate::next_stack_item_id(), items))
         }
         5 => {
             let len = cursor.read_u32()? as usize;
@@ -297,7 +297,7 @@ fn decode_stack_value_depth(cursor: &mut Cursor<'_>, depth: usize) -> Result<Sta
                 items.push(decode_stack_value_depth(cursor, depth + 1)?);
             }
             Ok(StackValue::Struct(
-                crate::next_stack_item_id() as u64,
+                crate::next_stack_item_id(),
                 items,
             ))
         }
@@ -312,14 +312,14 @@ fn decode_stack_value_depth(cursor: &mut Cursor<'_>, depth: usize) -> Result<Sta
                 let value = decode_stack_value_depth(cursor, depth + 1)?;
                 items.push((key, value));
             }
-            Ok(StackValue::Map(crate::next_stack_item_id() as u64, items))
+            Ok(StackValue::Map(crate::next_stack_item_id(), items))
         }
         7 => Ok(StackValue::Interop(cursor.read_u64()?)),
         8 => Ok(StackValue::Iterator(cursor.read_u64()?)),
         9 => Ok(StackValue::Null),
         10 => Ok(StackValue::Pointer(cursor.read_i64()?)),
         11 => Ok(StackValue::Buffer(
-            crate::next_stack_item_id() as u64,
+            crate::next_stack_item_id(),
             cursor.read_bytes()?.to_vec(),
         )),
         _ => Err("invalid stack value tag".to_string()),
